+++ src/index.js
@@ -13,7 +13,7 @@ import Request from './request';
 import RequestList from './request_list';
 import { openRequestQueue } from './request_queue';
 import SettingsRotator from './settings_rotator';
-import { apifyClient, getMemoryInfo, isProduction } from './utils';
+import { apifyClient, getMemoryInfo, isProduction, isAtHome } from './utils';
 import { browse, launchWebDriver } from './webdriver';
 
 /* globals module */
@@ -44,6 +44,8 @@ module.exports = {
     readyFreddy,
     getMemoryInfo,
     getApifyProxyUrl,
+    isAtHome,
+    client: apifyClient,
 
     // Autoscaled pool
     AutoscaledPool,
@@ -77,9 +79,6 @@ module.exports = {
 
     // Settings rotator
     SettingsRotator,
-
-    // Utils
-    client: apifyClient,
 
     // Webdriver
     browse,